Renamed about us to history page. Still not sure about name. Add first draft of content. Still much content needed. Also layout still wonky. We need one tight page completed.
